Import an ngsfilter file

Import an ngsfilter file formatted for the OBITools package (Boyer et al. 2016).

read_ngsfilter(file, additional.sep = "=", ...)

Arguments

file

path/name of the input ngsfilter txt file

additional.sep

character indicating the field separator in the additional info column in the ngsfilter filter file.

...

other arguments to be pasted from read.table

Value

a data.frame

Details

The function read_ngsfilter creates a `pcrs` table from an ngsfilter file formatted for the OBITools package (Boyer et al. 2016).

References

Boyer, F., Mercier, C., Bonin, A., Le Bras, Y., Taberlet, P., & Coissac, E. (2016). obitools: a unix-inspired software package for DNA metabarcoding. Molecular Ecology Resources, 16(1), 176-182.

See also

Author

Lucie Zinger

Examples

# \donttest{

dir <- tempdir()
url = "https://raw.githubusercontent.com/metabaRfactory/metabaR_external_data/master/"

ngsfilter_file = "ngsfilter_GWM-768.new_2.txt"
ngsfilter_url = paste(url, ngsfilter_file, sep="")
ngsfilter_path <- file.path(dir, ngsfilter_file)
download.file(ngsfilter_url, ngsfilter_path)

ngsfilter <- read_ngsfilter(
              file = ngsfilter_path,
              sep = "\t",
              additional.sep = "=")

# }